Fix broken Terraform example
The previous version of this code failed with Terraform 0.12+ on an M1 Mac with the following error:

│ Error: Incompatible provider version
│
│ Provider registry.terraform.io/hashicorp/template v2.2.0 does not have a
│ package available for your current platform, darwin_arm64.
│
│ Provider releases are separate from Terraform CLI releases, so not all
│ providers are available for all platforms. Other versions of this provider
│ may have different platforms supported.

The issue is that the `template_file` directive is now deprecated and no download exists for darwin_amd64 for this provider.

I considered using the now current templatefile directive, but we don't have any files from which we're reading templates, so I used the more idiomatic variable interpolation.